leave suggestions in the suggestions channel (#10)
Instead of instantly deleting messages upon processing a new submission,
we add a reaction to it instead, and leave it in the channel. Once the
emoji gets denied or approved and moved to the public approval queue,
the message in the public suggestions channel gets deleted.

This has the advantage of letting the public know what blobs are
currently being processed.

A new column in the database schema had to be added in order to track
the message ID of the suggestion in the suggestions channel.

If the bot fails to delete the message or fails to find it at all, it
logs it in the bot log for staff to diagnose.

The "successful suggestion" message was amended with a tiny notice
encouraging blob artists to not delete their message in the suggestions
channel.