Increase test image GIF compatibility
This decreases the number of color codepoints in the image to 24, which
gives libimagequant a considerably higher number of free slots to
accommodate 'truer' color when compositing GIFs.

This also improves the efficiency and effectiveness of local/global
palette swapping, which while not usually increasing quality with PIL,
has a chance of decreasing output file size.